+++ scripts/webpack_append_data.js
@@ -3,4 +3,5 @@ if (window && !Html5QrcodeScanner) {
     var Html5QrcodeScanner = window.__Html5QrcodeLibrary__.Html5QrcodeScanner;
     var Html5Qrcode = window.__Html5QrcodeLibrary__.Html5Qrcode;
     var Html5QrcodeSupportedFormats = window.__Html5QrcodeLibrary__.Html5QrcodeSupportedFormats
+    var Html5QrcodeScannerState = window.__Html5QrcodeLibrary__.Html5QrcodeScannerState;
 }