+++ src/index.ts
@@ -15,3 +15,4 @@
 export { Html5Qrcode } from "./html5-qrcode";
 export { Html5QrcodeScanner } from "./html5-qrcode-scanner";
 export { Html5QrcodeSupportedFormats } from "./core";
+export { Html5QrcodeScannerState } from "./state-manager";